+++ upcloud/testdata/upcloud_server/server_metadata_s2.tf
@@ -1,0 +1,30 @@
+variable "prefix" {
+  default = "tf-acc-test-server-metadata-"
+  type    = string
+}
+
+variable "zone" {
+  default = "fi-hel2"
+  type    = string
+}
+
+resource "upcloud_storage" "this" {
+  title = "${var.prefix}storage"
+  zone  = var.zone
+  size  = 10
+  tier  = "maxiops"
+}
+
+resource "upcloud_server" "this" {
+  hostname = "${var.prefix}server"
+  zone     = var.zone
+  metadata = true
+
+  storage_devices {
+    storage = upcloud_storage.this.id
+  }
+
+  network_interface {
+    type = "utility"
+  }
+}