+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## Fixed
+
+- upcloud_server: correctly handle metadata toggling on resource update when template is not set.
+
 ## [5.23.2] - 2025-07-30
 
 ### Fixed